requirements.txt
flask~=1.1.2
numpy~=1.20.2
torch~=1.8.1+cu111
transformers~=4.5.1
pysaucenao~=1.5.0
discord.py
termcolor~=1.1.0
requests~=2.25.1
aitextgen~=0.5.2